+++ src/lib.rs
@@ -205,6 +205,12 @@
 //!
 //! <p></p>
 //!
+//! - `serde`:
+//!
+//!   Adds serialization support for `BumpBox` and vectors.
+//! 
+//! <p></p>
+//!
 //! - `nightly-allocator-api` *(requires nightly)*:
 //!
 //!   Enables `allocator-api2`'s `nightly` feature which makes it reexport the nightly allocator api instead of its own implementation.